Format string expected ints, but was receiving longs

+++ St. Olaf/ISO8601DateFormatter.m
@@ -795,7 +795,7 @@ static BOOL is_leap_year(NSUInteger year);
         if (offset == 0)
             str = [str stringByAppendingString:ISO_TIMEZONE_UTC_FORMAT];
         else
-            str = [str stringByAppendingFormat:ISO_TIMEZONE_OFFSET_FORMAT, offset / 60, offset % 60];
+            str = [str stringByAppendingFormat:ISO_TIMEZONE_OFFSET_FORMAT, (int)offset / 60, (int)offset % 60];
     }
 
     //Undo the change we made earlier
